♻️ int8 -> int

+++ app/types.go
@@ -13,9 +13,9 @@ type Pull struct {
 	State        string `json:"state"`
 	Merged       bool   `json:"merged"`
 	User         string `json:"user"`
-	Commits      int8   `json:"commits"`
-	Comments     int8   `json:"comments"`
-	Additions    int8   `json:"additions"`
-	Deletions    int8   `json:"deletions"`
-	ChangedFiles int8   `json:"changed_files"`
+	Commits      int    `json:"commits"`
+	Comments     int    `json:"comments"`
+	Additions    int    `json:"additions"`
+	Deletions    int    `json:"deletions"`
+	ChangedFiles int    `json:"changed_files"`
 }
